fix(clerk-js): Add a dummy base when constructing the url in isRedirectForFAPIInitiatedFlow
